Fixing tests to ensure Assertions and exceptions in inner threads fail tests
