+++ ambry-router/src/main/java/com.github.ambry.router/ProgressTracker.java
@@ -79,3 +79,40 @@ class ProgressTracker {
         || decryptionStatusTracker.hasSucceeded());
   }
 }
+
+/**
+ * Tracks decryption job status
+ */
+class DecryptionStatusTracker {
+  private boolean succeeded;
+  private boolean done;
+
+  /**
+   * Updates the decryption status as completed and succeeded
+   */
+  void setSucceeded() {
+    this.succeeded = true;
+    this.done = true;
+  }
+
+  /**
+   * Updates the decryption status as completed and failed
+   */
+  void setFailed() {
+    this.done = true;
+  }
+
+  /**
+   * @return {@code true} if the decryption is completed. {@code false} otherwise
+   */
+  boolean isDone() {
+    return done;
+  }
+
+  /**
+   * @return {@code true} if the decryption has succeeded. {@code false} if failed.
+   */
+  boolean hasSucceeded() {
+    return succeeded;
+  }
+}
